Refactor sample application to improve modularity and semantics extraction
- Extract UI components from `SampleActivity.kt` into dedicated files: `LoginScreens.kt`, `FeedScreens.kt`, `FormScreens.kt`, and `SampleComponents.kt`.
- Implement `SampleSemanticsExtractor.kt` to handle `A11yNode` extraction, including viewport filtering and Compose view traversal.
- Simplify `SampleActivity.kt` by using modularized screen components and a new `SampleScreen` data model for navigation.
- Improve accessibility of "Fixed" sample screens by adding proper semantic roles, content descriptions, and touch target sizes.
- Add `SampleViewportTag` and `BrokenSampleContentTag` to facilitate targeted accessibility scanning within the sample app.
